+++ wdtk-datamodel/src/main/java/org/wikidata/wdtk/datamodel/json/jackson/JacksonObjectFactory.java
@@ -27,8 +27,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import org.wikidata.wdtk.datamodel.helpers.DatamodelConverter;
 import org.wikidata.wdtk.datamodel.helpers.Datamodel;
+import org.wikidata.wdtk.datamodel.helpers.DatamodelConverter;
 import org.wikidata.wdtk.datamodel.interfaces.Claim;
 import org.wikidata.wdtk.datamodel.interfaces.DataObjectFactory;
 import org.wikidata.wdtk.datamodel.interfaces.DatatypeIdValue;
@@ -62,6 +62,7 @@ import org.wikidata.wdtk.datamodel.json.jackson.datavalues.JacksonValue;
 import org.wikidata.wdtk.datamodel.json.jackson.datavalues.JacksonValueGlobeCoordinates;
 import org.wikidata.wdtk.datamodel.json.jackson.datavalues.JacksonValueItemId;
 import org.wikidata.wdtk.datamodel.json.jackson.datavalues.JacksonValueMonolingualText;
+import org.wikidata.wdtk.datamodel.json.jackson.datavalues.JacksonValuePropertyId;
 import org.wikidata.wdtk.datamodel.json.jackson.datavalues.JacksonValueQuantity;
 import org.wikidata.wdtk.datamodel.json.jackson.datavalues.JacksonValueString;
 import org.wikidata.wdtk.datamodel.json.jackson.datavalues.JacksonValueTime;
@@ -82,7 +83,8 @@ public class JacksonObjectFactory implements DataObjectFactory {
 	public ItemIdValue getItemIdValue(String id, String siteIri) {
 		if (id.length() > 0 && id.charAt(0) == 'Q') {
 			Integer numericId = Integer.valueOf(id.substring(1));
-			JacksonInnerEntityId innerEntity = new JacksonInnerEntityId(
+			JacksonInnerEntityId innerEntity;
+			innerEntity = new JacksonInnerEntityId(
 					JacksonInnerEntityId.JSON_ENTITY_TYPE_ITEM, numericId);
 
 			JacksonValueItemId result = new JacksonValueItemId();
@@ -96,8 +98,19 @@ public class JacksonObjectFactory implements DataObjectFactory {
 
 	@Override
 	public PropertyIdValue getPropertyIdValue(String id, String siteIri) {
-		// Jackson has no dedicated property id values:
-		return Datamodel.makePropertyIdValue(id, siteIri);
+		if (id.length() > 0 && id.charAt(0) == 'P') {
+			Integer numericId = Integer.valueOf(id.substring(1));
+			JacksonInnerEntityId innerEntity;
+			innerEntity = new JacksonInnerEntityId(
+					JacksonInnerEntityId.JSON_ENTITY_TYPE_PROPERTY, numericId);
+
+			JacksonValuePropertyId result = new JacksonValuePropertyId();
+			result.setValue(innerEntity);
+			result.setParentDocument(getParentItemDocument("Qunknown", siteIri));
+			return result;
+		} else {
+			throw new IllegalArgumentException("Illegal property id: " + id);
+		}
 	}
 
 	@Override
@@ -121,9 +134,10 @@ public class JacksonObjectFactory implements DataObjectFactory {
 	public GlobeCoordinatesValue getGlobeCoordinatesValue(long latitude,
 			long longitude, long precision, String globeIri) {
 		JacksonInnerGlobeCoordinates innerCoordinates = new JacksonInnerGlobeCoordinates(
-				((double)latitude / GlobeCoordinatesValue.PREC_DEGREE), 
-				((double)longitude / GlobeCoordinatesValue.PREC_DEGREE), 
-				((double)precision / GlobeCoordinatesValue.PREC_DEGREE), globeIri);
+				((double) latitude / GlobeCoordinatesValue.PREC_DEGREE),
+				((double) longitude / GlobeCoordinatesValue.PREC_DEGREE),
+				((double) precision / GlobeCoordinatesValue.PREC_DEGREE),
+				globeIri);
 		JacksonValueGlobeCoordinates result = new JacksonValueGlobeCoordinates();
 		result.setValue(innerCoordinates);
 		return result;
@@ -237,8 +251,8 @@ public class JacksonObjectFactory implements DataObjectFactory {
 		if (claim.getMainSnak() instanceof JacksonSnak) {
 			result.setMainsnak((JacksonSnak) claim.getMainSnak());
 		} else {
-			result.setMainsnak((JacksonSnak) dataModelConverter
-					.copySnak(claim.getMainSnak()));
+			result.setMainsnak((JacksonSnak) dataModelConverter.copySnak(claim
+					.getMainSnak()));
 		}
 
 		Map<String, List<JacksonSnak>> qualifiers = new HashMap<>();
@@ -312,9 +326,19 @@ public class JacksonObjectFactory implements DataObjectFactory {
 			List<MonolingualTextValue> labels,
 			List<MonolingualTextValue> descriptions,
 			List<MonolingualTextValue> aliases, DatatypeIdValue datatypeId) {
+		return getPropertyDocument(propertyId, labels, descriptions, aliases,
+				Collections.<StatementGroup> emptyList(), datatypeId);
+	}
+
+	@Override
+	public PropertyDocument getPropertyDocument(PropertyIdValue propertyId,
+			List<MonolingualTextValue> labels,
+			List<MonolingualTextValue> descriptions,
+			List<MonolingualTextValue> aliases,
+			List<StatementGroup> statementGroups, DatatypeIdValue datatypeId) {
 		JacksonPropertyDocument result = new JacksonPropertyDocument();
-		initializeTermedDocument(result, propertyId, labels, descriptions,
-				aliases);
+		initializeTermedStatementDocument(result, propertyId, labels,
+				descriptions, aliases, statementGroups);
 
 		switch (datatypeId.getIri()) {
 		case DatatypeIdValue.DT_ITEM:
@@ -354,27 +378,8 @@ public class JacksonObjectFactory implements DataObjectFactory {
 			List<StatementGroup> statementGroups,
 			Map<String, SiteLink> siteLinks) {
 		JacksonItemDocument result = new JacksonItemDocument();
-		initializeTermedDocument(result, itemIdValue, labels, descriptions,
-				aliases);
-
-		Map<String, List<JacksonStatement>> jacksonStatements = new HashMap<>();
-		for (StatementGroup sg : statementGroups) {
-			String propertyId = sg.getProperty().getId();
-			List<JacksonStatement> propertyStatements = new ArrayList<>(sg
-					.getStatements().size());
-			jacksonStatements.put(propertyId, propertyStatements);
-
-			for (Statement s : sg) {
-				if (s instanceof JacksonStatement) {
-					propertyStatements.add((JacksonStatement) s);
-				} else {
-					propertyStatements
-							.add((JacksonStatement) this.dataModelConverter
-									.copy(s));
-				}
-			}
-		}
-		result.setJsonClaims(jacksonStatements);
+		initializeTermedStatementDocument(result, itemIdValue, labels,
+				descriptions, aliases, statementGroups);
 
 		Map<String, JacksonSiteLink> jacksonSiteLinks = new HashMap<>(
 				siteLinks.size());
@@ -393,10 +398,12 @@ public class JacksonObjectFactory implements DataObjectFactory {
 		return result;
 	}
 
-	private void initializeTermedDocument(JacksonTermedDocument document,
+	private void initializeTermedStatementDocument(
+			JacksonTermedStatementDocument document,
 			EntityIdValue entityIdValue, List<MonolingualTextValue> labels,
 			List<MonolingualTextValue> descriptions,
-			List<MonolingualTextValue> aliases) {
+			List<MonolingualTextValue> aliases,
+			List<StatementGroup> statementGroups) {
 
 		document.setJsonId(entityIdValue.getId());
 		document.setSiteIri(entityIdValue.getSiteIri());
@@ -415,6 +422,25 @@ public class JacksonObjectFactory implements DataObjectFactory {
 
 		document.setLabels(buildTermMapFromTermList(labels));
 		document.setDescriptions(buildTermMapFromTermList(descriptions));
+
+		Map<String, List<JacksonStatement>> jacksonStatements = new HashMap<>();
+		for (StatementGroup sg : statementGroups) {
+			String propertyId = sg.getProperty().getId();
+			List<JacksonStatement> propertyStatements = new ArrayList<>(sg
+					.getStatements().size());
+			jacksonStatements.put(propertyId, propertyStatements);
+
+			for (Statement s : sg) {
+				if (s instanceof JacksonStatement) {
+					propertyStatements.add((JacksonStatement) s);
+				} else {
+					propertyStatements
+							.add((JacksonStatement) this.dataModelConverter
+									.copy(s));
+				}
+			}
+		}
+		document.setJsonClaims(jacksonStatements);
 	}
 
 	private Map<String, JacksonMonolingualTextValue> buildTermMapFromTermList(