+++ wdtk-datamodel/src/main/java/org/wikidata/wdtk/datamodel/json/jackson/datavalues/JacksonValueString.java
@@ -8,6 +8,8 @@ import org.wikidata.wdtk.datamodel.interfaces.ValueVisitor;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.databind.JsonDeserializer.None;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 
 /*
  * #%L
@@ -36,6 +38,7 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
  *
  */
 @JsonIgnoreProperties(ignoreUnknown = true)
+@JsonDeserialize(using = None.class)
 public class JacksonValueString extends JacksonValue implements StringValue {
 
 	private String value;