+++ wdtk-datamodel/src/test/java/org/wikidata/wdtk/datamodel/json/jackson/JsonTestData.java
@@ -37,6 +37,7 @@ import org.wikidata.wdtk.datamodel.json.jackson.datavalues.JacksonValue;
 import org.wikidata.wdtk.datamodel.json.jackson.datavalues.JacksonValueGlobeCoordinates;
 import org.wikidata.wdtk.datamodel.json.jackson.datavalues.JacksonValueItemId;
 import org.wikidata.wdtk.datamodel.json.jackson.datavalues.JacksonValueMonolingualText;
+import org.wikidata.wdtk.datamodel.json.jackson.datavalues.JacksonValuePropertyId;
 import org.wikidata.wdtk.datamodel.json.jackson.datavalues.JacksonValueQuantity;
 import org.wikidata.wdtk.datamodel.json.jackson.datavalues.JacksonValueString;
 import org.wikidata.wdtk.datamodel.json.jackson.datavalues.JacksonValueTime;
@@ -52,14 +53,13 @@ public class JsonTestData {
 
 	public static final DataObjectFactory JACKSON_OBJECT_FACTORY = new JacksonObjectFactory();
 
-	// TODO maybe decompose the time a bit to have less magic strings in it
+	// TODO maybe decompose the time a bit to have fewer magic strings in it
 
 	public static final String JSON_ENTITY_TYPE_ITEM = "item";
+	public static final String JSON_ENTITY_TYPE_PROPERTY = "property";
 
 	// the id's used in the tests
 	public static final String TEST_PROPERTY_ID = "P1";
-	public static final PropertyIdValue TEST_PROPERTY_ID_VALUE = Datamodel
-			.makeWikidataPropertyIdValue(TEST_PROPERTY_ID);
 	public static final String TEST_ITEM_ID = "Q1";
 	public static final int TEST_NUMERIC_ID = 1;
 	public static final String TEST_STATEMENT_ID = "statement_foobar";
@@ -71,9 +71,13 @@ public class JsonTestData {
 	// stand-alone descriptions of Value-parts
 	public static final String JSON_STRING_VALUE = "{\"type\":\""
 			+ JacksonValue.JSON_VALUE_TYPE_STRING + "\",\"value\":\"foobar\"}";
-	public static final String JSON_ENTITY_ID_VALUE = "{\"type\":\""
+	public static final String JSON_ITEM_ID_VALUE = "{\"type\":\""
 			+ JacksonValue.JSON_VALUE_TYPE_ENTITY_ID
 			+ "\",\"value\":{\"entity-type\":\"" + JSON_ENTITY_TYPE_ITEM
+			+ "\",\"numeric-id\":" + TEST_NUMERIC_ID + "}}";
+	public static final String JSON_PROPERTY_ID_VALUE = "{\"type\":\""
+			+ JacksonValue.JSON_VALUE_TYPE_ENTITY_ID
+			+ "\",\"value\":{\"entity-type\":\"" + JSON_ENTITY_TYPE_PROPERTY
 			+ "\",\"numeric-id\":" + TEST_NUMERIC_ID + "}}";
 	public static final String JSON_TIME_VALUE = "{\"type\":\""
 			+ JacksonValue.JSON_VALUE_TYPE_TIME
@@ -130,8 +134,10 @@ public class JsonTestData {
 
 	public static final JacksonValueString TEST_STRING_VALUE = (JacksonValueString) JACKSON_OBJECT_FACTORY
 			.getStringValue("foobar");
-	public static final JacksonValueItemId TEST_ENTITY_ID_VALUE = (JacksonValueItemId) JACKSON_OBJECT_FACTORY
+	public static final JacksonValueItemId TEST_ITEM_ID_VALUE = (JacksonValueItemId) JACKSON_OBJECT_FACTORY
 			.getItemIdValue("Q1", Datamodel.SITE_WIKIDATA);
+	public static final JacksonValuePropertyId TEST_PROPERTY_ID_VALUE = (JacksonValuePropertyId) JACKSON_OBJECT_FACTORY
+			.getPropertyIdValue("P1", Datamodel.SITE_WIKIDATA);
 	public static final JacksonValueTime TEST_TIME_VALUE = (JacksonValueTime) JACKSON_OBJECT_FACTORY
 			.getTimeValue(2013, (byte) 10, (byte) 28, (byte) 0, (byte) 0,
 					(byte) 0, (byte) 11, 0, 0, 0, TimeValue.CM_GREGORIAN_PRO);